Show user the download costs. Some improvements.
Show user the job output files download cost.
Get the job output details via /jobs/' + job.id + '/outputFiles?projection=detailedFtepFile'.
Fix job wms layers toggling.
When a new item is created, select it (cannot expect it to be the last one).
Ask confirmation when deleting a service.
Set the jobs filter order to same as it is for others.

Change-Id: I9a885a14c74f76354df6bf84fe3c840a01387fcd