Optimise job orchestration, eagerly release gRPC response observer
Refactor the f-tep-orchestrator to remove the stored reference to each
gRPC response observer for a job. This should allow gRPC connections to
close, and moves the orchestrator towards a more stateless model.

The FtepJobLauncher clients are now responsible for tracking job
completion independently, if required, through a new JobDataService
service.

Change-Id: I201120f102fb48566340b28e49b371d2190c4ac1